pin hp rpms at a specific version
This is a separate commit because we may regret the pin.
It depends on how frequently HP upgrades the packages and
whether the older versions are still available in the repo.

For now it's nice to see - in the source code - what version
of the packages we expect to install.